수정/삭제 테스트 완료

+++ src/main/java/com/backend/gini/domain/repository/BoardRepository.java
@@ -4,5 +4,5 @@ import com.backend.gini.domain.entity.BoardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
 }
